+++ linux-0.11-lab/0.11/kernel/fork.c
@@ -137,7 +137,7 @@ int copy_process(int nr, long ebp, long edi, long esi, long gs, long none,
 	set_tss_desc(gdt + (nr << 1) + FIRST_TSS_ENTRY, &(p->tss));
 	set_ldt_desc(gdt + (nr << 1) + FIRST_LDT_ENTRY, &(p->ldt));
 	p->state = TASK_RUNNING; /* do this last, just in case */
-	return last_pid;
+	return (int)last_pid;
 }
 
 int find_empty_process(void)
